refactor (move directory) : dto 구조 변경
