+++ naver-api-executor/src/main/java/me/ohtaeg/application/MovieService.java
@@ -1,9 +1,8 @@
 package me.ohtaeg.application;
 
-import me.ohtaeg.api.dto.SearchWord;
+import me.ohtaeg.api.dto.request.SearchWord;
 import me.ohtaeg.domain.repository.SearchRepository;
 import me.ohtaeg.domain.response.Movie;
-import me.ohtaeg.domain.response.SearchApi;
 import org.springframework.stereotype.Service;
 
 @Service
