+++ naver-api-executor/src/main/java/me/ohtaeg/application/SearchService.java
@@ -1,6 +1,6 @@
 package me.ohtaeg.application;
 
-import me.ohtaeg.api.dto.SearchWord;
+import me.ohtaeg.api.dto.request.SearchWord;
 import me.ohtaeg.domain.response.SearchApi;
 
 public interface SearchService {
