+++ naver-api-executor/src/main/java/me/ohtaeg/domain/response/Blog.java
@@ -1,7 +1,7 @@
 package me.ohtaeg.domain.response;
 
-import me.ohtaeg.api.dto.BlogRequest;
-import me.ohtaeg.api.dto.SearchWord;
+import me.ohtaeg.api.dto.request.BlogRequest;
+import me.ohtaeg.api.dto.request.SearchWord;
 import me.ohtaeg.domain.response.constant.RequestVariableName;
 import me.ohtaeg.util.PropertyUtils;
 import org.springframework.util.CollectionUtils;
