+++ naver-api-executor/src/main/java/me/ohtaeg/domain/response/SearchApi.java
@@ -1,9 +1,7 @@
 package me.ohtaeg.domain.response;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import me.ohtaeg.api.dto.SearchWord;
-
-import java.util.List;
+import me.ohtaeg.api.dto.request.SearchWord;
 
 public abstract class SearchApi {
     @JsonIgnore
